不強制更新相簿的index.html

+++ py/wuliu-photo-album.py
@@ -137,8 +137,9 @@ def write_pics_msgp(pics: dict, album_info: dict, album_path: Path):
     
     src = Path(Webpages).joinpath('index.html')
     dst = album_path.joinpath('index.html')
-    print(f'Write => {dst}')
-    shutil.copyfile(src, dst)
+    if not dst.exists():
+        print(f'Write => {dst}')
+        shutil.copyfile(src, dst)
 
 
 def create_album(pics: list, album_info: dict, album_path: Path, thumb_size):
